added OrthogonalCamera

+++ Cameras/PerspectiveCamera.cpp
@@ -1,7 +1,7 @@
 #include "PerspectiveCamera.h"
 
-Ray PerspectiveCamera::getRayAtPixel(double x, double y) const
+Ray PerspectiveCamera::getRayAtPixel(float x, float y) const
 {
-    Vector3d dir = m_rightDir * (x - m_width/2.) + m_upDir * (-y + m_height/2.) + m_eyeDir * m_focalLength;
+    Vector3f dir = m_rightDir * (x-m_width/2.) + m_upDir * (-y+m_height/2.) + m_eyeDir * m_focalLength;
     return {m_pos, dir.normalized()};
 }
